form changes
- redid upload functionality to go straight to cloud storage from the frontend. this is a) faster b) allows for  larger files to be uploaded
- added splash screen to form page so we can query data on load
- reworked image url functionality to use persistent download URLs (won't expire)
- fixed toast and crop styling
- added loading overlay while image processing and application uploading occurs
- added check for users who have already submitted applications
NOTE: this check is purely in frontend and is not secure as the state could be manipulated by a malicious user who wants to apply multiple times (but idk why anyone would waste their time doing that). I may secure this from the cloud function on backend if i have time
- added auth check on form page

NOTE: the pages for non-authed users and already-applied users look awful because i'm tight on time. i might make a prettier version later when I get a free moment